upstream(core): Remove the direct dependency from `iota-core` to `iota-move-build` (#7998)
- Upstream range: [v1.40.3, v1.41.2)
- Port commit:
-
https://github.com/MystenLabs/sui/commit/e54c57db0ca21a7546ed7d14be57a35492f9fc4b
- Description:
Previously, `iota-move-build` is linked into `iota-core` as a normal
dependency. The only reason for this is that some of the `iota-core`
testing utilities depend on `iota-move-build` and some other crates
depend on the `iota-core` testing utilities. This PR cleans this up a
bit
so that the testing infrastructure that depends on `iota-move-build` is
not re-exported, and therefore `iota-move-build` only needs to be a
dev-dependency of `iota-core`.

Specifically, the `iota::test_utils` code that is only used
inside of `iota-core` was split off into `iota::unit_test_utils`, and
gated
this file behind `#[cfg(test)]`. The tiny bit of shared
testing code that uses `iota-move-build` was moved into
`iota_move_build::test_utils`.

Similarly,
`iota_core::authority::authority_test_utils` into
`iota_core::authority::auth_unit_test_utils`
were split.

Part of #6149

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes
