+++ crates/iota-cluster-test/src/test_case/fullnode_build_publish_transaction_test.rs
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use async_trait::async_trait;
-use iota_core::test_utils::compile_basics_package;
 use iota_json_rpc_types::IotaTransactionBlockEffectsAPI;
+use iota_move_build::test_utils::compile_basics_package;
 use iota_types::{base_types::ObjectID, object::Owner};
 use jsonrpsee::rpc_params;
