+++ crates/iota-core/src/unit_tests/execution_driver_tests.rs
@@ -45,10 +45,8 @@ use crate::{
     },
     safe_client::SafeClient,
     test_authority_clients::LocalAuthorityClient,
-    test_utils::{
-        init_local_authorities, init_local_authorities_with_overload_thresholds,
-        make_transfer_object_move_transaction, make_transfer_object_transaction,
-    },
+    test_utils::{make_transfer_object_move_transaction, make_transfer_object_transaction},
+    unit_test_utils::{init_local_authorities, init_local_authorities_with_overload_thresholds},
 };
 
 #[expect(dead_code)]
